Remove stale TODO regarding member rule combination for -if rule.
The scenario we worried is that some of combinations of member rules are
equivalent if we're trying all combintaions of all sizes: kind of
range(memberKeepRules.size()). But we already limit the size of elements
per combination as the size of member rules to satisfy, and thus such
case does not happen at all.

Bug: 79486261
Change-Id: I0d05806fc2a803e7e2280fc46a5ca0c20faf7eea