Workaround Art instance-of verifier bug.
In rare cases the Art verifier will incorrectly propagate type
information from an instance-of back to the source of a move.

The problematic case is:

  move vA, vB
  instance-of vB, vA, Type
  if-eq/neq vB
  ...
  use at boolean type of vB

In this case, the type of vB will be seen as a reference type after
the if. Therefore, at the use at boolean type the verifier will
complain that vB has reference type and that boolean type was
expected.

This happens very rarely and this change works around it by
inserting a nop between the move and the instance-of instruction
which stops the type propagation from happening in the verifier.

R=zerny@google.com

Bug: 120985556
Change-Id: I02ec32cc1c4a1baa119548757ad2448dcf14dda5