Introduce escape analysis.
Escape analysis inputs IR code and the value of interest and checks if
the value could escape from the method, i.e., can be stored outside or
passed as parameters, etc.

------

Restrict name reflection optimization with help of escape analysis.
Now, we'll not apply the optimization to names that are used only inside
the method, because it's likely one-time name computation.
This can address most of regressions we saw before, but still not good
enough. Until it seems worthwhile, e.g., combined with Class#forName,
this optimization will be still off by default.

------

Also introduced another flag to enforce the optimization for testing.

Bug: 118536394, 120138731
Change-Id: I2b5e8b4fd832d5e38cf8433219009502cb54ee18