Workaround another mediatek Lollipop VM bug.
Mediatek has a VM that cannot deal with an unreachable empty
loop... So, we cannot use that to workaround a Dalvik tracing
compiler bug where the compiler can run off the end of the
instruction stream. Changing the workaround to generate a return
instead of an empty loop should deal with both issues.

R=zerny@google.com

Bug: 117907456, 119895393
Change-Id: I1fdcaf2cf651f83b88567839d402e67c799ce095