Allow storing arbitrary class types in fields with an interface type
This CL updates the type checker to accept field-assignments where an object of class type C is assigned to a field that has interface type I and C is not a subtype of I. This is needed since interface types are treated like Object according to the JVM specification (https://docs.oracle.com/javase/specs/jvms/se9/html/jvms-4.html#jvms-4.10.1.2-100). Prior to this CL, the body of methods with such instructions were replaced by "throw null".

Note that, although R8 should generally not encounter such field-assignments, we have seen such code in libraries that have been shrunk with another compiler than R8.

With this CL, R8 is technically still unsound. Due to assignment rules in the JVM specification, it is not safe to trust that a value that is being read from a field with interface type I actually has type I. This CL additionally extends the `InvalidTypeTest` to demonstrate that all of DX, D8, Proguard, and R8 are unsound in this regard.

Bug: 120277396
Change-Id: Ic403d769d26adab3a5cc3e30cb7e9f24e2cd41dd