Collect non-null after invoke for all arguments including receiver
Before this change the non-null after invoke was only correct for
static calls, as the receiver argument was skipped when collecting,
and not offset when applying.

Changed to include the receiver (which is trivially non-null after the
invoke) to avoid having to offset the index when used.

Added tests for direct, virtula and interface calls.

Bug: 71500340
Change-Id: I1a8af990d21d9501f77ef1b579e9e9ebf1b10e67