Generalize types used by tests to super-type / interface
Some of our unit tests test specific behavior where implementation
specific types are used, such as ImmutableList. This creates a
dependency on com.google.common which require relocating the r8 tests
as well.

This change updates most of these dependencies except for BiMap in
google.common which is also exposed to the tests.

Change-Id: I16443dca19a7fa9b1991bf06fb14dad5bd64684d