Replace nullable bit in TypeLatticeElement with NullLatticeElement.
Primitive types are always not null, so detaching it from type lattice.
Rather, we can move it down to ReferenceLatticeElement, which requires
another refactoring of canonicalization of reference types.
Instead of dual relations, we need three variants per type:
MAYBE_NULL, DEFINITELY_NOT_NULL, and DEFINITELY_NULL. In factory,
the MAYBE_NULL variant is registered, and then the other two variants
are linked on demand.

Bug: 72693244
Change-Id: I9d6d035f145506c8eea98f60c82b6805f695680d