Redundant const number removal
This CL introduces an optimization that exploits that we can sometimes learn the constant value of an SSA value that flows into an if-eq of if-neq instruction.

Consider the following example:

 1. if (obj != null) {
 2.  return doStuff();
 3. }
 4. return null;

Since the we know that `obj` is null in all blocks that are dominated by the false-target of the if-instruction in line 1, we can safely replace the null-constant in line 4 by `obj`, and thereby save a const-number instruction.

Bug: 120257211
Change-Id: Id7a9b9647046649686041b1c24a93dc07b89383b