Forward GraphLense to IRBuilder
In the IRBuilder we assert that the given method has been changed to have void-return type according to the graph lens, whenever we change a non-void return instruction to a void return instruction. Prior to this CL this assertion would always fail along this code path because the identity graph lense is being passed.

Bug: 123317382
Change-Id: Ib24070b6c3a532e4dfd792c4c0d4d10faaf922b1