Fix ProguardMapReader assertion when we have synthesized methods
These are generated by class merging. We insert an extra method to ensure super behaviour.
The extra inserted method maps back to the original method, in addition
to the actual method that also maps back.

The only way we can currently see that this is synthesized by the class merger
is that there is no line numbers for these.

Bug: 120894450
Change-Id: I584af27e50cb45f2b0c7dcbccedcab7e5ab79221