Disable assertion in vertical class merger
This assertion fails for the Instabug-Android test app. The failure is harmless though: the assertion is failing because a non-abstract class does not implement all the methods from the interface it implements. This generally does not happen for input that has been generated by javac, but it may happen if the program has been optimized in advance.

Change-Id: Icfb774773bf018ca65eda598d67c7d805d90d1d6