Disallow direct manipulation of method arrays
This CL changes the API of DexClass, such that usages can no longer manipulate the arrays DexClass.directMethods and DexClass.virtualMethods directly. To achieve this, the methods DexClass.directMethods() and DexClass.virtualMethods() no longer return a direct reference to the underlying arrays. Instead, these arrays need to be mutated via appropriate methods on DexClass.

The motivation for this encapsulation is to be able to easily introduce an assertion that will fail upon the first change to DexClass.directMethods or DexClass.virtualMethods that introduce a duplicate method definition.

Bug: 123730537
Change-Id: Ia9459453391fe9a52b04b692b1ce277dfa7f27c7