Collect dependencies of all kept annotations and add with reason
All references for all annotations that are kept should be added to
live types, otherwise we will see errors when combined with
-keepattributes Exceptions.

There should be no need to keep library annotations and mark
references as live types for library methods, classes and fields.

It is a bit weird that throwing is treated as an annotation in our
shared representation - refactoring this have a feature request here:
b/124218160.

Further, when marking a type as live, we should give a reason for
keeping it such that we can investigate live-type relationships in the
graph inspector: b/124217402.

Bug: 124019003
Change-Id: I1b877eb5c4adedd662c4f629bfe17b484e3fa1f5