Introduce AppView in D8
This CL introduces AppView also for D8. The AppView for D8 is created via AppView.createForD8(), whereas the one for R8 is created via AppView.createForR8(). The only difference between the two is the result of appView.enableWholeProgramOptimizations().

This CL also introduces a new interface DexDefinitionsSupplier, which requires the presence of the definitionFor() methods. Both AppInfo and AppView now implements this interface. This way, it is possible to pass either an AppInfo or AppView instance to places where the AppInfo or AppView is not needed for anything else than retrieving definitions.

Note that this CL does *not* ensure that all uses of AppInfo always use the most recent AppInfo instance. Additional cleanup is needed to get to this stage.

Bug: 114469298, 127783942
Change-Id: Id76c64ec18885a43249696509da316d9c85ee35d