Add two stages of error handling for applymapping
1)
Add handling of clashing sources and targets directly in the mapping
file before trying to apply the map on the source.

2)
Check for accidently having clashes of existing classes not present in
the map. This error will happen if one takes application source
together with a map and use in a test source:

Application:
class B { ... }

Mapping:
B -> A

Test:
class A { ... }

We also check for clashing of methods and fields by piggybacking on
the existing infrastructure that tries to give fresh names to methods
and fields. If a fresh-name could not be given from the map, it must
be because of a clash.

Bug: 126503704
Bug: 123092153
Bug: 121305642
Change-Id: I5e4d6d553cbce634d2b0ae262a1091a2dc390560