MethodNameMinifier reserve methods when using applymapping
When producing an optimized library we can do optimizations that are
not visible during second compilation in studio, since the default is
no optimizations and no shrinking. Example:

interface A {
  void foo();
}

interface B {
  void foo();
}

class C implements A {
  void bas() { ... }
  void foo() { ... }
}

class D implements B {
  void foo() { ... }
}

Names for interfaces are collected in one go across all interfaces
such that interfaces methods that have the same name will get the same
renaming. However, if A.foo() is never used and no other users of
A.foo() exists, we can de-virtualize it and move it to C. We then
could have a map as follows:

A      -> A: (no methods because of de-virtualization)
B.foo  -> c
C.bas  -> c
C.foo  -> d

When we now run without optimizations (second pass) we will not
de-virtualize and assign the name c to C.foo, which obviously collides
with B.foo.

This CL fixes this issue by consulting the strategy for a given map
before reserving the name.

Bug: 126503704
Bug: 123092153
Bug: 121305642
Change-Id: I1c3a915f8eca664893b8021c956b58e8d78278c0