Refactor on-demand link on ReferenceTypeLatticeElement
This is a precursor to fix not widening when narrowing.

The current implementation bakes in duality by having primary and
variant as on-demand fields. The current nullability lattice have
three values and we will add Bottom shortly, so it will be hard to
maintain the fields and have on-demand links back and forth.

The on-demand links is changed to be a shared pointer struct. This
encapsulates creating and linking all similar reference lattice
elements. This makes the synchronicity at bit easier to digest as well.

Bug: 125492155
Change-Id: Iada100551f2865ee23684e9d56a775e75a049e81