Include all methods in vertical class merger's collision analysis
Prior to this CL the vertical class merger only performed its collision detection on the set of targeted methods. This needs to the set of all program methods.

Note that, the set of targeted methods is generally close to the set of all program methods, except when tree shaking is disabled.

Bug: 129570925
Change-Id: I008d19f86c9e844acef49a77f5f44a1d985eeb46