Extend member pool collection to library types
The member pool collection currently does not consider library types. This means that the member pool collection may erroneously report that a given signature is not present in the hierarchy, although the signature is actually declared in a super type that is defined in the library.

In addition to extending the member pool collection to library types, this also adds a test that we do not remove unused arguments from virtual methods when that would lead to a collision with a library method.

Bug: 129794426, 65810338
Change-Id: Idf5a514c8b6fe9fffecad94c29e48caa571388d1