Rewrite of ServiceLoader.load to NewInstance when possible
Users using unreleased versions of R8 in kotlin projects can observe
different behaviour between versions as a result of using
ServiceLoader.load for reading kotlin metadata. This is because
multiple service loaders for a service will now exist on the class
path.

This CL will eliminate the ServiceLoader.load in R8lib by rewriting it
to the known instances.

Bug: 120436373
Bug: 129240946
Change-Id: Ie96925e2c3c5e70f2c81f32b9f8eff71b6d7ac91