Throw null if an instruction will do.
Based on SSA value's nullability as well as uninstantiated type info,
if an instruction requires non-null receiver, e.g., instance-{get|put},
we can replace it with `throw null` when the receiver is known to be
null.

One caveat is, for field instructions, resolution-related errors come
first. This CL refactors dead condition of {instance|static}-get and
rather simulates field resolutions. If it will raise errors other
than NPE, we should honor that.

Bug: 72443802, 129530569
Change-Id: Iecddc7feafd07da8496634835ddc0cacc08feea8