Handle annotated unused argument removal with missing parameter annotations
Code generated by certain versions of javac can have a mismatch between the
number of parameter annotations and the number of parameters. This is known
to happen for non-static inner classes, where the synthetic argument to
<init> for the outer instance is not counted.

The normal procedure is to keep the structure with this mismatch through to
the output. However, when removing arguments the rewriting now takes this
into account, and will explicitly add missing parameter annotation
information.

Change-Id: I6ae62a5040f18454ccb23cdc4f719a5208311561