Allow for using minification with applymapping
This will add a test ensuring that all classes and members that do not
have an explicit mapping will be minified depending on them being kept
or not.

The internal state of MethodNamingState that previously kept track of
reserved names and new name indices are now split in much the same way
as the field minification state. If this is not done, the assertion
that a child will only use indicies greater than its parent will not
hold.

Bug: 130736358
Change-Id: Ia7b9eef1a9ae7237f2ff17d9f1aef39a3c968a51