Rewrite renaming strategies to ask for and specify pre-reserved names
The R name is always reserved in a package, which at the very least
will cause an infinite loop if trying to name the same thing.

This CL rewrites the minification strategies to generate a fresh-name
by asking a passed in predicate if the new-name is used or not. We put
in an additional assert that checks the new name is not used.

Bug: 133646663
Change-Id: Iee86a7fea8945446253c17ed25970ac1ef8d378e