Fix type of outline arguments
Use the already calculated type lattice element of the values flowing into
the outline to determine the outline argument type. The previous calculation
based on subtypes only did not work in all situations.

Fixed equals of oultines as it was missing the outline argument types. The
fix with the correct argument types revealed this.

Bug: 133215941
Change-Id: I02d980f9dacb57adabe96c1ec8e840fcdcf5b871