Don't inline methods that returns an integer as a boolean on dalvik.
Dalvik will do a typecheck and not verify the code if we do.

This changes the inliner to disallow inlining a method that returns a
boolean if we can't be sure that the value returned is a boolean type.
Kotlin will generate integer fields that it returns as boolean.
This only crashes on dalvik.

Bug: 134304597
Change-Id: I3b4cb8d3fbfe4dc87c64f5c948ca66f9577feb6d