Migrate String and Objects backport/desugar tests
This changes the harness to look at method invocations in all methods instead of just main. Some of the tests were getting large so splitting them up visually is a readability win.

Because all methods are being checked now and we sometimes need to refer to other methods on the target type, an annotation is added to ignore any invocations inside that method.

Test: tools/test.py --dex_vm all --no-internal -v *Backport*
Change-Id: I9f75f4a216da729521470a5ca3104be0bc0969a7