Rework InterfaceMethodNameMinifier to support applymapping
Applymapping uses the supplied map as an initial seed, in practice,
reserving names for all classes, interfaces and members. The current
implementation did not support having a reserved name that was not the
name itself.

The algorithm is basically the same as before, however it is much
faster now due to caching. Compiling YT 13.17 would use around 10.5
seconds in the InterfaceMethodNameMinifier before, now it is
around 0.5 seconds (430 ms).

Bug: 122924648,131532229
Change-Id: I448e10a109110f80800e820f9c623b2b485acefd