Mark widening overriding methods as targeted if also on interface
We have to ensure not to remove methods that are widening access and
also defined on interface. This CL extends the Enqueuer to mark such
methods as targeted. This is done by building up the scoped method set
according to the type hierarchy and for each widening function check
if it is also defined on one of the interfaces the class implements.

There are 13 such instances widening methods in GMS-core and all of
them was not removed before thus this CL will have very little
negative size impact.

One could do without the actual check on the interfaces to save some
runtime performance. On GMScore there is around 37 K of methods that
will now be targeted and the size increase is 2600 bytes.

Bug: 136698023
Change-Id: Ic2e0176904c109b4eb20b662f29f67fac088f9f9