Refine constant-array analysis
This CL refines the utility that determines if an array is guaranteed to be constant through-out the current method.

If a constant array has been found in the code, the analysis checks if any of the following instructions may modify the array. This is currently done by checking if any of the subsequent instructions may have side effects. This CL updates the analysis to instead check if any of the subsequent instructions may trigger another method to be invoked, which is more precise.

Bug: 137041585
Change-Id: I55af40d8205062f289f0ead9b9b5b168cd1bf31b