Ensure member classes are not non-member classes.
For anonymous classes (non-null outer and null inner-name), we can
recover by erasing outer-class references.

For local classes (non-null inner-name and enclosing member), though,
it is unclear whether the intention was a member class or local class.
So we can't remove outer-class references in this case, and rather fail
hard.

Bug: 137881258
Change-Id: I4f1159fe978729bc72d959575b3707172ab2edb0