Record static edge from the context of static invokes to the static receiver.
This deals with one of the two issues in b/135474075, where a singleton pattern
(the type Absent) was reported to have only a cyclic reason for being kept. A
regression test is added for the remaining issue: the super type (the type
Optional in the report) is reported as not being kept despite being in the
output.

Bug: 124501298
Bug: 135474075
Change-Id: I7c4f08e8f7a16e364b72b577a2c8154ad97f82a8