Add applymapping test requiring renaming of kept classes members
This is basically the default setup for tests that use any kind of
mocking. We should support the most basic renamings even when
something is marked as kept in the test (program path).

The loop-hole regarding applymapping only being a naming seed is, that
the renamings are done in the library and it would be strange to not
apply them on the program path.

Change-Id: Id7f5a6ea308b8c89a61c0fa7b3b0859eb65bb77b