Rewrite minification to only rename if targets can be renamed
This will make the applymapping minification be consistent with being
a naming seed in the straight forward case.

For testing purposes it will not work if we did not allow renaming on
library and classpath. To have this fit in with keep rules and naming
the change in semantics is that keep-rules actually only apply to
program classes.

This change therefore allows changing members of kept program classes
if the renaming of those members has been done for library or
class-path classes.

Bug: 135173935
Change-Id: I354d04e860dbc944e356da97097961296b7a34b8