Reproduce b/139432507: double-inliner misses simple inlining candidates.
If a certain method is double-inlining candidate (by calling a
double-inlining target) but itself is an inlining candidate as a single
caller, that method is inlined into the caller while its body still has
potentially inlinable method call.

Double-inliner, later, reprocesses those double callers, which has been
already inlined into the single caller. It should either reprocess that
single caller or allow double-inlining target to be inlined in the
beginning.

Bug: 139432507
Change-Id: I4005bb5bc499f9e0444c00e7af3b63c3fb722307