Move optimization info structures to same package
Currently, the mutator methods on UpdatableMethodOptimizationInfo are public. Thus, it is currently possible to mutate a method optimization info object *directly*, circumventing the optimization feedback:

  method.getMutableOptimizationInfo().markMayNotHaveSideEffects()

This can potentially lead to subtle bugs, because the changes made directly to the optimization info could be overwritten when the delayed optimization info mechanism updates all the method optimization infos when a wave ends.

Moving all the optimization info structures into the same package is a first step towards fixing this problem, since this allows us to make all the mutator methods package-private, such that they can only be accessed via a optimization info feedback instance.


Besides, this CL makes a few minor changes to the optimization info:
* removes the UpdatableMethodOptimizationInfo info, which is only implemented by MethodOptimizationInfoImpl,
* renames MethodOptimizationInfoImpl to UpdatableMethodOptimizationInfo,
* renames DefaultMethodOptimizationInfoImpl to DefaultMethodOptimizationInfo.

Bug: 140214568
Change-Id: I9652bf70021827ffdb3864e888afe7be5b6f2a6b