Allow space characters in SimpleName for the DEX format 040.
We allow both ASCII and Unicode spaces (everything in the
Unicode category 'Zs'). This includes characters:

  0x20, 0xa0, 0x1680, 0x2000..0x200a, 0x202f, 0x205f, 0x3000

Since this feature is not available in old versions of dex
format, we need conditional name validation that depends on
the Android API level. This validation is done on reading
.class and .dex files and on writing .dex files. In-between
reading and writing relaxed validitiy checks are used that
allow spaces regardless of the API level

Change-Id: I377b6998e58b091566b9db70c93b10453d68ce2e