Lookup -assumevalues when trying to replace a call with a constant.
We overweigh -assumenosideeffects. But, if the matched rule doesn't have
a return value, while there is another rule to specify assumed value,
each rule interfere each other.

While keeping the precedence as-is, we need to switch to use -assumevalues
if -assumenosideeffects without return value is about to be used for
attempting to find a replacement constant.

Bug: 141127471
Change-Id: Icd73958acf5c25c6caaf2fffecbcfc34f8ec09f8