Introduce Assume variant with no assumptions.
With this variant, we can aggressively add aliases and see if existing
optimizations were designed with aliases in mind.

Currently, its insertion and removal are aligned with other assume
instructions, and updated optmizations in between are as follows:
* String(Builder)Optimizer have not passed updated nullability beyond
aliases.
* Some places put affected values *after* replacing the current
instruciton, which already changed the users!
* Added more check points to make sure IR code is consistent w.r.t SSA
and associated type lattices are verifiable.

Bug: 120920488
Change-Id: I33f22cc5c8666140650c1932396a3a48d0d9381c