Update non-null propagation for static fields
This CL removes the part of ClassInitializerDefaultsOptimization that updates the optimization info of fields, since this does strictly not have anything to do with replacing static-put instructions by encoded dex values.

Instead, this is now being done in a new FieldValueAnalysis, which can be extended to instance fields in follow up work.

Change-Id: I0afac5f371da62e64ef734f1baf2a93a831793e2