Extend the range of assume instructions, part 4.
Outliner is updated to embrace assum instructions while looking for
outlining candidate instructions in a row. Its logic on widening the
window of instructions regards assume instructions same as constants,
i.e., without increment, and then it will not add assume instructions
into the outline candidate code, which is also same as how constants are
handled.

Bug: 120920488
Change-Id: Idee99290b6ed29639ffefd8bf1e5aacb7ec89b85