Preparation for adding -whyareyounotinlining support
This CL is in preparation for adding support for the new Proguard testing directive -whyareyounotinlining. For example, this CL introduces a WhyAreYouNotInliningReporter, and passes an instance of this type around to the methods that cause us to not inline a method.

As a first step, this CL also adds support for reporting why an invoke that does not have a single target is not inlined. A corresponding test is added.

Bug: 142108662
Change-Id: I6f9ccca2cc9d95caaa0bade072401747f063423a