Add test for running benchmark apps in +/- 10 % with expectation
This will enable the hard-failing tests to ensure we are staying in
bounds memory wise when compiling apps with R8.

For each benchmark app, we will try to compile with what was found to
be OOM limit + 10 % to ensure compilation still
succeeds. Additionally, we will test compiling with OOM limit - 10 %
with the expectation of getting an OOM. The test will fail if we do
not get one, to let us know that the OOM limit has to change.

Bug: 121018500
Change-Id: Id99f670da5271a16809da016394dd37ac061448c