Heuristic for not inlining methods with monitor instructions
When inlining a method that has at least one monitor instruction into a method that has at least one monitor instruction, the heuristic will prevent inlining if the number of values that are used in a monitor-enter instruction will exceed 4 after inlining.

The heuristic considers two lock values to be the same if they are identical (ignoring aliases introduced by assume instructions) or if they refer to the same const class (since such values should become identical after constant canonicalization).

The motivation for this is that the register allocator will pin a register for each monitor-enter value. Thus, if too many values are used in monitor-enter instructions, this will lead to high register pressure.

Change-Id: I7e9e25996aa55c11c40312f8e2847d300da11dc4