Disable vertical merging if base has dependency outside main dex
This CL will add an additional check in the vertical class merger when
merging bases with constructors if compiling with a main dex list.
The check will look through constructors for references outside the
main dex and disable merging if a dependency is found.

The check requires scanning the code of constructors so the check is
placed as the last one.

Bug: 142909854
Change-Id: I8d046f7666290fc8e358b28c3f030e4c3093df04