Spell out the different uses of initializer info
Prior to this CL, an instance initializer was classified as 'trivial' if it only assigns arguments into instance fields on the enclosing class.

This fact is used in several different places, exploiting different characteristics about such 'trivial' initializers (e.g., the receiver never escapes, it does not have any external side effects).

The problem with this design is that it hinders generalizing the notion of 'trivial' initializers (for example, to generalize the class inliner), since this may violate the expectations that we have to a 'trivial' initializer in some places.

To this end, this CL removes the notion of 'trivial' initializers, and instead introduces an InitializerInfo object, which can be queried for each of the distinct pieces of information about an instance initializer that we rely on.

Change-Id: I74ca8ed29577a0f30f4d88750e94e9bb17cc68ba