Do not retain spuriously reachable proto messages
The proto schemas that are encoded in each dynamicMethod() may reference other proto messages if the proto has a non-primitive field. The problem with this is that the proto schemas cannot be optimized until after the second round of tree shaking. Therefore, we want to avoid that the second round of tree shaking traces through the const-class instructions that are in each dynamicMethod() at the time of the second round of tree shaking.

This is achieved by introducing a new use registry, ProtoEnqueuerUseRegistry, which does not trace the const-class instructions in dynamicMethod().

Then, when the fixpoint is reached, we manually trace the const-class instructions that will remain after the proto schema has been optimized. This is done by the ProtoEnqueuerExtension.

Change-Id: If426d5170fefcb6821c89c47f477814055d89d37