Consider all implemented interfaces when removing unused arguments
To ensure that it is possible to remove unused arguments for a virtual
method all overriding methods are considered both above and below in the
hierachy. This check was missing interfaces which could be implemented
on classes below in the hierachy.

This is conservative, as it is not checked if a given method is
actually the only implementation of an interface method.

Bug: 143934502
Change-Id: Ie02f66544b4d437e41e67f6c7a85087206dda9a7