Improve handling of register allocation for intervals with same start.
In particular, allocate registers to intervals with hints first. Also,
allocate registers to intervals in the order of their hints. This to
avoid allocating a register to an interval that is a hint for another
interval.

This avoids a bunch of unnecessary aliases in code. We have seen
customized lollipop VMs with bugs that crash dex2oat if there are
aliases of parameters. This change removes the aliasing for the
particular case that we know about.

R=christofferqa@google.com, sgjesse@google.com

Bug: 142682636
Change-Id: Iee77f3f525831e11827cbacfdf2a7d5c55976dc3