Add synthetic bridges to missing interface methods implementations
When specifying a method keep rule on a class, that is a missing
implementation of an interface method, we now insert synthetic bridges
to ensure invoke-interface will continue to work when called on the
class.

This is a more precise implementation than just keeping the hierarchy,
as it is done for compat mode.

Bug: 143590191
Change-Id: Ic82f9592e0bc83b0500144cc7a8244e321719d50