Change ClassInlinerEligibility.returnsReceiver to OptionalBool
Prior to this CL, `returnsReceiver` was set to true if there was a return instruction that returned the receiver, even if another return instruction in the same method did not return the receiver. In addition to this, the previous implementation did not account for the fact that a method can return the receiver indirectly through an alias.

This CL uses the newly introduced OptionalBool to avoid that the class inliner's cost analysis is optimistic.

Bug: 144825216
Change-Id: I7bb51aa1d1af8477f667ee143131ff6848860157