Improve the speed of lowering constants for binop/lit8 and binop/lit16
The computational complexity is only there when checking if a binop
instruction can fit in the binop/2addr format, in which case lowering
the constant should not be done.

The main change is to collect the non constant in-values to the binop
instructions, and then find the last user of these values in the block
being analyzed to speed up checking for later use in the same block.

Local per-block instruction numbering is introduced to speed up checking
if one instruction is before another.

Finally an ArrayList was changed to a HashSet, as it was only used to
check containment.

This brings the compilation time of the supplied repro down from unknown
to less than a minute (30 seconds for d8).

Bug: 145052564
Change-Id: Ifde8f941f9347d60961717146a7435a68a6e4888