Remove ClassInitializerInfo
This removes the need for ClassInitializerInfo by leveraging the fact that we already compute dynamic lower bound type information for static final singleton fields.

A prerequisite for class inlining singleton fields is that the corresponding static-get instruction cannot trigger observable side effects. Therefore, this CL optimizes the methods of synthesized lambda classes and publishes the gathered optimization feedback before starting a wave. This ensures that we will mark the class initializers of lambdas as not having observable side effects, such that we can class inline the INSTANCE fields of these lambdas.

(This was not previously needed because the class inliner implicitly assumed that these class initializers did not have observable side effects.)

Bug: 145642327, 120814598
Change-Id: If72c86442c69d5d1d23fead04dc69768aaa7e189