Add SourceDebugExtensionRewriter to support kotlin inline functions
Kotlin inline functions are "appended" to the end of source files,
which makes debugging exceptions in inlined functions hard since the
point to non-existing code.

This CL adds support for synthesizing debug positions by parsing the
SourceDebugExtension attribute for finding positions and files, and
matching it with the local variable table for finding method names.

Bug: 141817471
Change-Id: I5b5d6b18e24d800db45050f32dca5d9b49275d60