Support "$JDK_HOME" as argument to --lib for command line tools
When running d8 and r8 through the CLI allow the argument to --lib to
be a directory of a JDK (the JAVA_HOME location for that JDK). In that
case add the JDK runtime classes as library classes.

This is the only way to get the runtime classes for JDK 9+, as they do
not ship with a rt.jar anymore.

Bug: 143335486
Change-Id: Idacdfc9ba85cef70a50255d064a4bdb9ff115288