Retrace: Align the regexp parser with the StackTrace parser
Before this CL using either of the parsers would yield a different
result. Most notably when having no line-number information and a
source, the output would be stripped from the regexp parser.

Also, when dealing with ambiguous cases, the StackTrace parser would
print the result in a "nicer" way.

This CL aligns the behavior between the two such that they will print
the same for most inputs. One thing to note here is that the ambiguous
format will now just append <OR> before the first non-whitespace, for
the simple reason that the output will then be consistent.

Bug: 145356709
Change-Id: I49e58e04e95f8ccaa80ddf7a57ce5bdc486d93f7